file restructure and added todos and fixmes
