+++ scripts/axostop.sh
